Project fix

+++ index.php
@@ -1,9 +1,11 @@
 <?php
 
 
-use models\Company;
-use models\Customer;
 use eftec\bladeone\BladeOne;
+use helper\Admin;
+use models\Customer;
+
+require_once "config.php";
 //    require_once "models/Company.php";
 //    require_once "models/Conversation.php";
 //    require_once "models/Customer.php";
@@ -11,30 +13,123 @@ use eftec\bladeone\BladeOne;
 // vietoi 3 require_once 1 vendor failas naudojant composer
 
 
-require_once "vendor/autoload.php";
-//
-//foreach (Customer::all() as $customer ){
-//    echo  $customer->getName()."<br>";
-//}
-//foreach (Customer::get(1)->getConversations() as $conve){
-//    echo $conve->getDate();
+
+// -----------------------------------------------------------------------
+//                               LOGIN
+// ------------------------------------------------------------------------
+
+if(! isset($_SESSION['user'])){
+
+    header('location: login.php');
+    die();
+};
+
+//if(isset($_GET['logout'])){
+//    unset($_SESSION['vartotojas']);
+//    session_destroy();
+//   header('location: login.php');
+//    die();
 //}
 
-        foreach (Company::all() as $company){
-            echo $company->getName()."<br>";
-        }
+//if(isset($_POST['atsijungti'])){
+//    session_destroy();
+//    header("location: login.php");
+//}
+
+// -----------------------------------------------------------------------
+//                               LOGIN
+// ------------------------------------------------------------------------
+
 
         if(isset($_GET['delete'])){
             Customer::get($_GET['delete'])->delete();
             header('location: index.php');
+
         }
-//foreach (Customer::get(1)->getCompany() as $comp){
-//    echo $comp->getName();
+
+
+
+
+
+// -----------------------------------------------------------------------
+//                               DELETE
+// ------------------------------------------------------------------------
+
+
+
+// -----------------------------------------------------------------------
+//                             Gražina prisijungusi vartotoja
+// ------------------------------------------------------------------------
+ //$vartotojas=Admin::getUser($_SESSION['vartotojas'][1]);
+
+
+// -----------------------------------------------------------------------
+//                               Gražina prisijungusi vartotoja
+// ------------------------------------------------------------------------
+
+
+
+
+// -----------------------------------------------------------------------
+//                               PRISIJUNGIMAS
+// ------------------------------------------------------------------------
+
+
+ // $vartotojas2= Admin::login('dominykas@gmail.com','dominykas');
+  // echo $vartotojas2;
+
+
+// -----------------------------------------------------------------------
+//                              PRISIJUNGIMAS
+// ------------------------------------------------------------------------
+
+
+// -----------------------------------------------------------------------
+//                               SKIRTINGI NAV BARAI
+// ------------------------------------------------------------------------
+
+//foreach ($_SESSION['vartotojas'] as $info){
+//    if($info == 'SuperAdmin'){
+//       $user = new SuperAdmin;
+//    }else{
+//        $user = new Admin;
+//    }
+//
 //}
-   echo Customer::get(1)->getCompany()->getName();
+
+
+
+// -----------------------------------------------------------------------
+//                               SKIRTINGI NAV BARAI
+// ------------------------------------------------------------------------
+
+
+// -----------------------------------------------------------------------
+//                               TIKRINAM AR PRISIJUNGE
+// ------------------------------------------------------------------------
+
+
+
+//        if(Admin::isLogin()){
+//            echo "Prisijunge";
+//
+//        }else{
+//            echo "Neprisijunge";
+//}
+
+
+
+// -----------------------------------------------------------------------
+//                                    TIKRINAM AR PRISIJUNGE
+// ------------------------------------------------------------------------
+
+
+
+
+
     $blade = new BladeOne( __DIR__."/views",__DIR__."/compile",BladeOne::MODE_DEBUG);
     echo $blade->run("index");
-        ?>
+   ?>
